app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==44.0.1
jinja2==3.1.2
python-multipart==0.0.6
pytz==2023.3
requests
pandas
python-dotenv
sentence-transformers
faiss-cpu
numpy
